The atoi(3) function is deprecated: The ahttp://www.gnu.org/software/libc/manual/html_mono/libc.html#index-atoi-2567 It is recommended to replace it with strtol(3), which detects errors, as opposed to the former. Since strtol returns long, this also modifies the data types of some variables. This however is not a problem, because a long integer fits anywhere an int does.

+++ src/include/request.h
@@ -197,7 +197,7 @@ struct session_request
     int user_home;              /* user_home request(VAR_ON/VAR_OFF) */
     
     /*-Connection-*/
-    int port;
+    long port;
     /*------------*/
     
     /* file descriptors */
